# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Global AI Missions API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Sets the httpOnly session cookie. Rate limited per (client IP, email).",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "parameters": [
          {
            "name": "gam_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gam Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/missions/{mission_id}/attempts": {
      "post": {
        "tags": [
          "attempts"
        ],
        "summary": "Start Attempt",
        "description": "Start the mission, or return the OPEN attempt (200) so a finished run can be submitted.",
        "operationId": "start_attempt_api_missions__mission_id__attempts_post",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          },
          {
            "name": "gam_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gam Session"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateView"
                }
              }
            }
          },
          "200": {
            "description": "The open attempt already existed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/attempts/{attempt_id}": {
      "get": {
        "tags": [
          "attempts"
        ],
        "summary": "Get Attempt",
        "description": "The current StateView: how a student resumes after a reload or a lost connection.",
        "operationId": "get_attempt_api_attempts__attempt_id__get",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          },
          {
            "name": "gam_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gam Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/attempts/{attempt_id}/advance": {
      "post": {
        "tags": [
          "attempts"
        ],
        "summary": "Advance",
        "description": "CONTINUE on a narrative or consequence node; node_id must be the current node.",
        "operationId": "advance_api_attempts__attempt_id__advance_post",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          },
          {
            "name": "gam_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gam Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/attempts/{attempt_id}/answer": {
      "post": {
        "tags": [
          "attempts"
        ],
        "summary": "Answer",
        "description": "Lock the checkpoint's answer; the server grades it. No outcome field (F-01).",
        "operationId": "answer_api_attempts__attempt_id__answer_post",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          },
          {
            "name": "gam_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gam Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/attempts/{attempt_id}/submit": {
      "post": {
        "tags": [
          "attempts"
        ],
        "summary": "Submit",
        "description": "Two-phase and idempotent: grade once, then Maya's feedback. Returns the Report.",
        "operationId": "submit_api_attempts__attempt_id__submit_post",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          },
          {
            "name": "gam_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gam Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/attempts/{attempt_id}/report": {
      "get": {
        "tags": [
          "attempts"
        ],
        "summary": "Get Report",
        "operationId": "get_report_api_attempts__attempt_id__report_get",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          },
          {
            "name": "gam_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gam Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdvanceRequest": {
        "properties": {
          "node_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Node Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "AdvanceRequest"
      },
      "AnswerRequest": {
        "properties": {
          "node_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Node Id"
          },
          "option_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-z0-9]{1,8}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Option Id"
          },
          "text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "AnswerRequest"
      },
      "AnswerResponse": {
        "properties": {
          "maya_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maya Line"
          },
          "state": {
            "$ref": "#/components/schemas/StateView"
          }
        },
        "type": "object",
        "required": [
          "maya_line",
          "state"
        ],
        "title": "AnswerResponse",
        "description": "No outcome / is_correct field (F-01): the story carries what happened."
      },
      "AttemptRecord": {
        "properties": {
          "attempt_number": {
            "type": "integer",
            "title": "Attempt Number"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "submitted_at": {
            "type": "string",
            "title": "Submitted At"
          },
          "ending": {
            "$ref": "#/components/schemas/EndingRef"
          },
          "story_minutes_used": {
            "type": "integer",
            "title": "Story Minutes Used"
          },
          "rescue_used": {
            "type": "boolean",
            "title": "Rescue Used"
          },
          "hints_received": {
            "type": "integer",
            "title": "Hints Received"
          }
        },
        "type": "object",
        "required": [
          "attempt_number",
          "started_at",
          "submitted_at",
          "ending",
          "story_minutes_used",
          "rescue_used",
          "hints_received"
        ],
        "title": "AttemptRecord"
      },
...
      "Clock": {
        "properties": {
          "time": {
            "type": "string",
            "title": "Time"
          },
          "minutes_left": {
            "type": "integer",
            "title": "Minutes Left"
          },
          "train_departed": {
            "type": "boolean",
            "title": "Train Departed"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "time",
          "minutes_left",
          "train_departed",
          "label"
        ],
        "title": "Clock"
      },
...
      "DiaryEntryView": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "clock": {
            "type": "string",
            "title": "Clock"
          },
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "checkpoint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiaryCheckpoint"
              },
              {
                "type": "null"
              }
            ]
          },
          "maya_decision": {
            "type": "string",
            "enum": [
              "quiet",
              "hint",
              "rescue"
            ],
            "title": "Maya Decision"
          }
        },
        "type": "object",
        "required": [
          "seq",
          "clock",
          "node_id",
          "kind",
          "location",
          "text",
          "checkpoint",
          "maya_decision"
        ],
        "title": "DiaryEntryView"
      },
...
      "FeedbackSource": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "fallback"
            ],
            "title": "Status"
          },
          "provider_label": {
            "type": "string",
            "title": "Provider Label"
          }
        },
        "type": "object",
        "required": [
          "status",
          "provider_label"
        ],
        "title": "FeedbackSource"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Interpretation": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "strength": {
            "type": "string",
            "enum": [
              "grammar",
              "vocabulary",
              "reading",
              "listening",
              "speaking"
            ],
            "title": "Strength"
          },
          "challenge": {
            "type": "string",
            "enum": [
              "grammar",
              "vocabulary",
              "reading",
              "listening",
              "speaking"
            ],
            "title": "Challenge"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "strength",
          "challenge",
          "recommendation"
        ],
        "title": "Interpretation"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "MayaView": {
        "properties": {
          "mood": {
            "type": "string",
            "enum": [
              "curious",
              "encouraging",
              "worried",
              "proud"
            ],
            "title": "Mood"
          },
          "decision": {
            "type": "string",
            "enum": [
              "quiet",
              "hint",
              "rescue"
            ],
            "title": "Decision"
          },
          "line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          }
        },
        "type": "object",
        "required": [
          "mood",
          "decision",
          "line"
        ],
        "title": "MayaView"
      },
      "MissedCheckpoint": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "multiple_choice",
              "fill_blank",
              "comprehension",
              "vocabulary"
            ],
            "title": "Type"
          },
          "skill": {
            "type": "string",
            "enum": [
              "grammar",
              "vocabulary",
              "reading",
              "listening",
              "speaking"
            ],
            "title": "Skill"
          },
          "cefr": {
            "type": "string",
            "enum": [
              "PRE_A1",
              "A1",
              "A2",
              "B1",
              "B2",
              "C1"
            ],
            "title": "Cefr"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "stimulus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AudioStimulus"
              },
              {
                "$ref": "#/components/schemas/TextStimulus"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stimulus"
          },
          "your_answer": {
            "type": "string",
            "title": "Your Answer"
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "maya_tip": {
            "type": "string",
            "title": "Maya Tip"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "type",
          "skill",
          "cefr",
          "prompt",
          "stimulus",
          "your_answer",
          "correct_answer",
          "explanation",
          "maya_tip"
        ],
        "title": "MissedCheckpoint"
      },
...
      "NextMission": {
        "properties": {
          "mission_id": {
            "type": "string",
            "title": "Mission Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "skill_focus": {
            "items": {
              "type": "string",
              "enum": [
                "grammar",
                "vocabulary",
                "reading",
                "listening",
                "speaking"
              ]
            },
            "type": "array",
            "title": "Skill Focus"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "state": {
            "type": "string",
            "enum": [
              "available",
              "in_progress",
              "waiting_to_submit",
              "completed",
              "locked",
              "in_preparation"
            ],
            "title": "State"
          },
          "unlock_hint": {
            "type": "string",
            "title": "Unlock Hint"
          }
        },
        "type": "object",
        "required": [
          "mission_id",
          "title",
          "skill_focus",
          "reason",
          "state",
          "unlock_hint"
        ],
        "title": "NextMission"
      },
      "NodeView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "narrative",
              "checkpoint",
              "consequence",
              "ending"
            ],
            "title": "Kind"
          },
          "scene": {
            "$ref": "#/components/schemas/SceneView"
          },
          "checkpoint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CheckpointView"
              },
              {
                "type": "null"
              }
            ]
          },
          "ending": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EndingRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "scene",
          "checkpoint",
          "ending"
        ],
        "title": "NodeView"
      },
...
      "Report": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "title": "Attempt Id"
          },
          "mission_id": {
            "type": "string",
            "title": "Mission Id"
          },
          "mission_title": {
            "type": "string",
            "title": "Mission Title"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "result": {
            "$ref": "#/components/schemas/ResultBlock"
          },
          "attempt_record": {
            "$ref": "#/components/schemas/AttemptRecord"
          },
          "interpretation": {
            "$ref": "#/components/schemas/Interpretation"
          },
          "feedback_source": {
            "$ref": "#/components/schemas/FeedbackSource"
          },
          "next_mission": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NextMission"
              },
              {
                "type": "null"
              }
            ]
          },
          "diary": {
            "items": {
              "$ref": "#/components/schemas/DiaryEntryView"
            },
            "type": "array",
            "title": "Diary"
          },
          "missed": {
            "items": {
              "$ref": "#/components/schemas/MissedCheckpoint"
            },
            "type": "array",
            "title": "Missed"
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "mission_id",
          "mission_title",
          "label",
          "result",
          "attempt_record",
          "interpretation",
          "feedback_source",
          "next_mission",
          "diary",
          "missed"
        ],
        "title": "Report"
      },
      "ResultBlock": {
        "properties": {
          "score_pct": {
            "type": "integer",
            "title": "Score Pct"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "incorrect": {
            "type": "integer",
            "title": "Incorrect"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillScore"
            },
            "type": "array",
            "title": "Skills"
          },
          "unmeasured_skills": {
            "items": {
              "type": "string",
              "enum": [
                "grammar",
                "vocabulary",
                "reading",
                "listening",
                "speaking"
              ]
            },
            "type": "array",
            "title": "Unmeasured Skills"
          },
          "suggested_level": {
            "$ref": "#/components/schemas/SuggestedLevel"
          }
        },
        "type": "object",
        "required": [
          "score_pct",
          "correct",
          "incorrect",
          "total",
          "skills",
          "unmeasured_skills",
          "suggested_level"
        ],
        "title": "ResultBlock"
      },
...
      "SceneView": {
        "properties": {
          "location": {
            "type": "string",
            "title": "Location"
          },
          "backdrop": {
            "type": "string",
            "title": "Backdrop"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/SceneLine"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "location",
          "backdrop",
          "lines"
        ],
        "title": "SceneView"
      },
...
      "StateView": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "title": "Attempt Id"
          },
          "mission_id": {
            "type": "string",
            "title": "Mission Id"
          },
          "mission_title": {
            "type": "string",
            "title": "Mission Title"
          },
          "status": {
            "type": "string",
            "enum": [
              "in_progress",
              "completed",
              "submitted"
            ],
            "title": "Status"
          },
          "clock": {
            "$ref": "#/components/schemas/Clock"
          },
          "node": {
            "$ref": "#/components/schemas/NodeView"
          },
          "maya": {
            "$ref": "#/components/schemas/MayaView"
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "mission_id",
          "mission_title",
          "status",
          "clock",
          "node",
          "maya"
        ],
        "title": "StateView"
      },
...
      "UserView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "type": "string",
            "enum": [
              "student",
              "teacher",
              "admin"
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "role"
        ],
        "title": "UserView"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}